Change CardTravel from JS to TS, refactor CardTravel styles
